requirements.txt
pandas>=1.5.0
numpy>=1.21.0
tqdm>=4.64.0
pyarrow>=10.0.0
python-dateutil>=2.8.0